+++ NEWS.md
@@ -1,8 +1,9 @@
 # TreeSearch 0.4.3.9004 (development)
 
  - `EasyTree()` 'shiny' graphical user interface for tree search.
+ - `SiteConcordance()` calculates exact site concordance
  - `Resample()` conducts bootstrap and jackknife resampling.
- - `Consistency()` calculates consistency and retention 'indices'
+ - `Consistency()` calculates consistency and retention 'indices'.
  - `MinimumLength()` and `MaximumLength()` calculate range of possible lengths
    of characters in a dataset on any tree
  - Set handling of 'gap' token (-) when creating Morphy object with `gap = `